fix(ci): cast SAAS_INVITE_DAYS/MAX_STAFF to int; make env-var test CI-safe
CI failed on the seeder commit (37321a5) with two real bugs, not an infra issue:

1. config('saas.invite_days') read via bare env(), so once a real .env file sets
   SAAS_INVITE_DAYS (both production's .env and CI's copied .env.example do),
   env() returns a numeric STRING rather than falling back to the coded int
   default. ProdDemoSeeder.php:153 passed that straight into addDays(), which
   Carbon 3's strict int|float typing rejects with a TypeError. StaffController
   already guarded this with an (int) cast at the call site -- the seeder didn't.
   Fixed at the root: cast in config/saas.php (max_staff too, same pattern,
   currently only safe by accident via its one call site), plus the seeder call
   site for defense in depth.

   This is why it passed locally before: the local .env doesn't set
   SAAS_INVITE_DAYS, so env() returned the literal int default. Verified the fix
   by reproducing the exact CI condition (SAAS_INVITE_DAYS set as a session env
   var) locally -- failed before the fix, passes after, on both SQLite and with
   the var unset.

2. Phase59SeederTest's superadmin-seeder test used putenv() to simulate
   SUPERADMIN_* env vars. Laravel's env() checks / before falling
   back to getenv(), and some CI PHP builds restrict putenv() -- it silently
   no-op'd on the GitHub Actions runner (seeder saw nothing, skipped, 0 users
   created) while working fine on Windows locally. Fixed by setting /
   directly in the test, which Laravel's repository checks first regardless of
   putenv() availability.

Verified: 472 passed on SQLite under both the normal local condition AND the
exact CI condition (string env var reproduced), plus a clean build.

+++ config/saas.php
@@ -16,8 +16,11 @@ return [
 
     // ST-Staff (S3): max users per store (owner + staff). Flat cap at launch;
     // becomes tier-based later via Tenant::seatLimit(). Invitations expire after N days.
-    'max_staff' => env('SAAS_MAX_STAFF', 5),
-    'invite_days' => env('SAAS_INVITE_DAYS', 7),
+    // Cast here (not just at call sites): once a real .env file sets these, bare
+    // env() returns a numeric STRING, which crashes Carbon's addDays() with a
+    // TypeError (int|float required) — bit ProdDemoSeeder in CI.
+    'max_staff' => (int) env('SAAS_MAX_STAFF', 5),
+    'invite_days' => (int) env('SAAS_INVITE_DAYS', 7),
 
     // ST-Legal (S6) / invoices: the registered business behind OSMS.
     'legal_entity' => env('SAAS_LEGAL_ENTITY', '[Your Registered Business Name]'),
